+++ examples/http-client/src/lib/models/http-response.model.ts
@@ -1,0 +1,9 @@
+import { HttpInternalRequestConfig } from './http-request-config.model';
+
+export interface HttpResponse<T = unknown, D = unknown> {
+  data: T;
+  status: number;
+  statusText: string;
+  headers: Headers;
+  config: HttpInternalRequestConfig<D>;
+}