+++ examples/http-client/src/lib/models/index.ts
@@ -1,0 +1,12 @@
+export type {
+  HttpInterceptorManager,
+  InterceptorId,
+  UseInterceptor,
+  FulfilledCallback,
+  RejectedCallback,
+  UseInterceptorParams,
+} from './http-interceptor-manager.model';
+export type { HttpAdapter } from './http-adapter.model';
+export type { HttpRequestConfig, HttpInternalRequestConfig } from './http-request-config.model';
+export type { HttpResponse } from './http-response.model';
+export { HttpError } from './http-error.model';